qcacmn: Pass chip ID to IPA
Currently, while passing chip id to IPA, there is
a check for mlo context, due to which initial soc
attachs are not able to set the chip id as mlo
conext gets set when all the socs are initialized.

Fix is not checking the mlo conext as its not
required untill all socs are initialized.

Change-Id: Ie80533ddbdcd4c2e3cc18eba153aae4ceaca10c4