qcacmn: Copy is_psd from SP master channel list to AFC channel list
The PSD support information of the country or regulatory domain is
lost for the channels in the current channel list that are not AFC
done but are supported by the regulatory(specifically for outdoor
deployment mode).

In outdoor deployment, for the channels that are not AFC done, the
channel information is copied from SP master channel list to AFC
channel list. While copying the "psd_flag" information is missed out.
While checking, if a country or regulatory domain supports PSD, the
"psd_flag" is read for the first channel in the current channel list
that is not disabled. Since in outdoor all the frequencies supported
by the regulatory are treated as not disabled,the "psd_flag" of the
first channel in the current channel list that is not disabled is read
and can be incorrect.

Copy the "psd_flag" information for the AFC channel list entries that
are not AFC done from the SP master channel list.

Change-Id: I2026dc053b95776d1c5028ad280a3ddee941d446
CRs-Fixed: 3713994