qcacmn: Fix API and reduce the debug level
Add check to see if the number of group in the MLO are valid and
then only do further sanity. For cases when the MLO group is not
yet initialized or is 0 return as false.
Also Reduce the debug level of prints so the console is not cluttered.

Change-Id: Ibd0c78e11fefe0aafdb34172c15251b8b89bb317
CRs-Fixed: 3843436
(cherry picked from commit 3eb82d35c74e833128c1b3f029447a8bd6f4a454)