qcacmn: Consider 0dbm as valid eirp
When the 6 GHz access point (AP) receives an Automatic
Frequency Control (AFC) response for a 320 MHz channel with
an EIRP of 0 dBm in the EIRP response or channel object, the
following conditions apply:

1) The best power mode is determined based on the highest
EIRP value.
2) If a specific power mode is not supported, the EIRP for
that mode is considered 0 dBm.
3) In scenarios where only the SP mode is available
(e.g., outdoor deployment), the best power mode defaults to
“REG_CURRENT_MAX_AP_TYPE (invalid power mode)" since the
EIRP is 0 dBm. Consequently, the AP will not come UP.

To fix the issue mentioned above, apply the following.
If the chosen best power mode is invalid, perform the
following checks:
i) Is the best power mode invalid and
ii) Is the deployment outdoors and
iii) Was an AFC power event received and
iv) Is the specified frequency and bandwidth present in the
AFC response and
v) Is the EIRP for the given frequency and bandwidth 0 dBm?

If all the above conditions are met, the best power mode is
SP with a power value of 0 dBm.
Therefore, consider 0 dBm in the AFC response as a valid
EIRP for selecting the best power mode.

Change-Id: Ib5c2bba2972b2378cfc353d1bf76df5f59b3eec1
CRs-Fixed: 3848257
(cherry picked from commit d601d7e75bfb54854ff46ba989a28ecce55d86d1)