qcacmn: Stop agile when there are no active vdevs
The intention is to stop the agile state machine
(send an agile abort command to the firmware and stop the preCAC timer
running in the host) when there are no active vdevs in the pdev.
This change established this behavior. However, change
I8a5bf3f2931f0c36a367dfebfc898ed2621f3516 modified this behavior by
canceling the preCAC timer if there are active vdevs in the pdev during
the process of vdev down causing the following issue:

When multiple AP vaps (e.g., vap1 and vap2) are brought up
on a DFS channel (e.g., channel 100) with the preCAC feature enabled,
the preCAC timer starts running. If vap2 is then brought down,
the tgt_vdev_mgr_down_send API is invoked during the vdev_down process,
which cancels the preCAC timer. Subsequently, when radar is injected on
the agile channel using the “radartool -i wifix bangradar 0 0 0 1”
command, the dfs_is_radarsource_agile() API fails because
the preCAC timer is not running. As a result, the agile radar is treated
as home channel radar, leading to a home channel change.

To fix this issue, do not cancel the agile CAC timer during every
VDEV DOWN. Instead, stop the agile state machine only if there
are no active vdevs in the pdev.

CRs-Fixed: 3943348
Change-Id: Ida71fe8fa35dec3dc9a19e351d4928631a1410fb