qcacmn: Add support for wifi radar on Marina
Marina chipset has a limitation where rx cannot be done
on chains adjacent to the tx chain. This limitations
needs to be taken care of while getting configs from
user. To achieve that, FW needs to send the data to host
where it signifies whether the chipset has any limitation
or not and host needs to parse the same and act
accordingly.

Change-Id: Ia601bba121fbdca8afa8b8ef25619bc4f9391b08
CRs-Fixed: 3968979