qcacmn: Optimize the Max Channel Switch Time value
In the current implementation, Max Channel Switch Time(MCST) is computed
using the below formula,

MCST value =
'100 milliseconds delay to receive the CSA completion event from the FW'
+ 'FW vdev restart time' + 'Host vdev restart' + 'CAC duration' +
'100 milliseconds delta time to send the beacon on the new channel'

Host vdev restart time is hardcoded as 520 milliseconds based on the max
number of vdevs in the host code.

Host vdev restart time =
'Time required to process the CSA completion event from the FW and to
send the multiple vdev restart command to the FW' +
'Time required process the restart response event from the FW and to
send the vdev up command to the FW for all the vdevs'.

For an AP device, add the changes to compute the host vdev restart time
based on the number of beaconing vdevs. For a single vdev, this value
will be around 100 milliseconds (considered the max host vdev restart
value of 2 GHz, 5 GHz and 6 GHz AP vdevs). From the second beaconing
vdev onwards 50 millisecond offset is added to the host vdev restart
time and the maximum host vdev restart time is limited to 520
milliseconds.

For a STA device, host vdev restart time is considered as 520
milliseconds to compute the channel switch time. STA uses this value to
check if the AP needs CAC on the new channel based on the received MCST
value from the AP.

Change-Id: Ib9c1d9ee9b8d830bc66a5e85e2bd11e04f278377
CRs-Fixed: 3971890