qcacmn: Add API to fetch width from opclass
Add regulatory core and dispatcher API to fetch channel width
and behav_limit from the opclass and the primary channel of the
ECSA IE.
The API (reg_get_chanwidth_and_behav_limit_from_opclass)invokes
"reg_get_opclass_from_map" function to fetch the opclass
table based on the "channel_map" global variable. This channel_map
variable is set according to the DFS domain received from the
firmware in the WMI_REG_CHAN_LIST_CC_EXT_EVENTID.
On the STA side, in the CSA from the Root AP, the
ECSA IE includes an opclass value. We first attempt
to find this opclass value in the opclass
table pointed to by channel_map. If it is not found there, we then
check if the opclass value is present in any opclass table (opclass_tbl).
This discrepancy can occur if the AP and the STA are configured with
different opclass table indices using the "cfg80211tool wifix
set_opclass_tbl "value" " command.

CRs-Fixed: 3986964
Change-Id: I0e614dd3b500cf402c5f6ca8c61ce9faed1bbbfd