qcacmn: API to get scan entry based on crypto
Add an API to get the scan entry based on the bssid
and the crypto params of the vdev.

When RSNO is being used, the highest supported RSN(O)
IE need not necessarily be the RSN parameter of the
connection. Therefore, always filter out based on the
crypto to get the proper security_info of the scan entry.

Change-Id: I6b9c632f6ec5a6c687f9fff63d0be7c39d49e3f5