qcacmn: Add reg opclass dipatchers
The channel spacing for opclasses that supports 80p80 is 80MHz and
can be confused with 80MHz opclasses. This may lead to incorrect
advertisement of supported opclasses.

The behavior limit for 80p80 opclass will be 80plus and can be used to
differentiate from 80MHz opclass. Change the visibility of the
function reg_is_opclass_entry_80p80() and create a dispatcher API
wlan_reg_is_opclass_entry_80p80(). This function checks the behavior
limit of the input opclass and determines if it 80p80 opclass or not.

Add a dispatcher to determine opclass table from the given country
ISO.

Change-Id: I6e4cb8852350e2533c3c6a149920f575f5f875f0
CRs-Fixed: 4011942