qcacmn: Fix hw_link_id_bitmap sent to FW on mode1 recovery
FW requires hw link id of the asserted non primary peers to skip syncing
for peer delete of that peer when primary peer is deleted.

With previous implementation, recovery link peers will be deleted
without waiting for FW responses, hence before the wmi peer delete of
the other peers could be formed, recovery link peer will be delted, and
hw link id bitmap will not be sent to FW at all.

To Fix the same, when the peers are deleted, will store the bitmap on
peer context, to be used for other peer deletes that follows.

Change-Id: Ia0b9454651d55a25c87159ab256253868231c333
CRs-Fixed: 4009157