cmn-dev: Remove nbuf NULL Assignment in comp2refill path
The local descriptor passed to PPEDS comp2refill API is a
union having first member as a next pointer and the second
member being the descriptor structure. When nbuf,
the first member is made NULL, it is making the
next pointer NULL as well since this is defined as a union.
As a result, the list is broken and the descriptors are
not added to the freelist, hence there is a descriptors
leak. This gerrit prevents the next pointer being set
to NULL.

Change-Id: I6a422aab02ee97c6cd09b98b8e59d8c58b0f09e7