qcacmn: Verify cfreq2 is within the primary channel's 320 MHz BW
When we trigger a CSA using the command cfg80211tool ath1 doth_ch_chwidth
73 5 320 3 127 0, with the primary frequency set to 73 and the 320 MHz
center frequency set to 127, the CSA is expected to fail because the
primary frequency 73 does not lie within the 320 MHz center frequency of
127. However, the CSA does not fail. This issue occurs because
reg_get_320_bonded_chan_array only checks if the user-provided cfreq2 is
valid, without verifying whether the primary frequency lies within the
user-configured cfreq2. To fix this, ensure that a valid bonding 320 MHz
pointer is found only if the primary frequency lies within the
user-configured 320 MHz center frequency.

Change-Id: I7ab78979f2e4b47cb36eeeba7447fbd09b6a2bdf
CRs-Fixed: 4022900