qcacmn: Add API to return number of connected peers
Add API to return the number of peers in connected state.

The existing API used in telemetry for of number of connected peers
also included the transient peers for which authenntication is ongoing
and may fail and might not end up in connected state.

This is requried for telemetry purposes where the number of
connected clients is required excluding any transient peers
created and should not be accounted until it is connected state.

Change-Id: Ifd4c86c82f14579c8979f6b95e17f4749ec940d9
CRs-Fixed: 4037977