qcacmn: Prohibit DFS puncture on non-11be channels
Ensure that the DFS puncture feature is not enabled on
non-11be channels by verifying if the DFS current channel
is 11be capable. If the user is operating in 11be mode and
then switches to 11axa mode, attempting to set the DFS
puncturing feature should be prohibited. This change
prevents the DFS puncture feature from being applied in
scenarios where it is not applicable.

CRs-Fixed: 4030894
Change-Id: I482ed66be6b3ea56da5e03b95af22ec497db22e8