qcacmn: Optimize DFS Random channel selection
When we bring up an AP on channel 100 HT160 with subchannel marking
enabled and then inject radar, the DFS random channel selection
algorithm occasionally selects a 20 MHz channel despite the
availability of 80 MHz channels. This issue arises because the algorithm
does not initially verify if the randomly chosen channel and bandwidth
are supported by the current regulatory domain. Instead, it selects a
random channel, queries the regulatory database on whether the
randomly chosen channel and input channel width is supported, and if
unsupported, reduces the bandwidth of the selected channel.

To address this issue, the random channel selection algorithm will
now only consider channels that are supported by the current
regulatory domain. This ensures that the optimal primary channel
and bandwidth combination is selected following radar detection.

CRs-Fixed: 4050107
Change-Id: I443b0324eb482fa51daa0f5e244a18d41c363ac8
(cherry picked from commit 9e8f31adc60ffa7944beb4139c2520da293ad490)