qca-cmn: Subchannel CAC before un-puncturing a user-punctured sub-channel
When radar is detected by the DFS module using the puncture/unpuncture
state machine (SM), the DFS module unpunctures a punctured channel after
the Non-Occupancy Period (NOP) and Channel Availability Check (CAC).

The new requirement specifies that when a user punctures a channel
(or configures a punctured channel or switches to a punctured channel),
the system should automatically attempt to complete CAC. If radar is
detected, it should complete NOP first and then CAC. This ensures that
when the user tries to unpuncture the channel in the future
(after NOP+CAC), it can be unpunctured without needing to CAC. Since
unpuncturing is a channel change operation and CAC during channel change
is done on full bandwidth, the stations connected to the Access Point(AP)
get disconnected. Once NOP and CAC are completed internally on a
punctured channel, it can be unpunctured without causing client
 disconnections.

Below are the changes done:
1] Added two new vdev variables
   is_user_punc     : Indicates whether the current puncture action was
                      carried out by the user or the radar.
   is_prev_user_punc: Indicates whether the last puncture action was
                      done by the user or the radar
   The values of these variables are used during the creation of DFS SM
   punctured Objects.

2] During vdev start, where we reinitialize the channel configuration
   and its puncture bitmap, It is necessary to identify who last punctured
   the channels in the bitmap and so we pass the value of 'is_prev_user_punc'
   which contains the info on whether the last puncture was done by the user
   or the radar.

3] New event "DFS_PUNC_SM_EV_USR_PUNC" added for the DFS SM to move the SM
   object to PUNCTURED state when user does the puncture.

4] New state "DFS_S_CAC_DONE" is added for the DFS SM for stopping the SM object from auto unpuncture.

Change-Id: I7667967425b6b0c44c8ea2c243c0cfd8aaba8af8
Signed-off-by: Ajithkumar Jayamoorthy <ajayamoo@qti.qualcomm.com>