qcacmn: Allow valid_link_bitmap to be reset during MLO SOC down sequence
The valid_link_bitmap for the number of active links in the MLO
manager is incremented alongside the num_links variable.

Remove the bitmap for the asserted SOCs when num_links is removed
to ensure that the two variables are synced together.

Change-Id: Iff69b190930e4e1a4c7c84f4a334f1e91d323ae7
(cherry picked from commit 5905e7317d284feb0d7db83963561f91a9fd8cda)