cmndev: update the soc_list based on static bypass flag
For the static bypass case, the soc id always starts from 0 even the
first SOC is bypassed. For example, if the SOC0 is bypassed, the
soc_id 0 will be used for SOC1 and soc_id 1 for SOC2 and so on.
Need to update the soc_list based on static bypass flag since for the
normal boot up or dynamic bypass time the soc_id will be intact.

To fix this issue, update the soc_list based on static bypass flag
by checking chip_id which is getting updated in
__qca_ol_update_mlo_capability.

Change-Id: I35c70ff62061777c14744921b5afa45e69784228
Signed-off-by: Madhavan Ganesan <quic_madhgane@quicinc.com>