qcacmn: Enhance DFS random channel selection
As per the current design of primary channel list in DFS
random channel selection, the 20 MHz channels that are
not part of the primary channel list are not considered for
random channel selection.

Thus, only the 20 MHz channels that are part of the primary
channel list are added for channel selection.

Consider a case where primary channel list is configured with
channels 52 and 106. The current operating channel is 106 and
80 MHz BW.

If we inject a radar in the operating BW, the successive random
channel selection has only one channel list to be used for
selection (channel 52). So it always downgrades to 20 MHz BW
and selects channel 52 in 20 MHz BW in the following
sequence:
  (1) Only the 20 MHz channels that are part of the primary
      channel list are added to available channel list and
      considered for random channel selection.
  (2) The target BW is 80 MHz, but the only available channel
      in the available channel list is 52, which is a 20 MHz
      channel.
  (3) The random channel selection expects all the subchannels
      to be available in the target BW (80 MHz) and be part
      of the available channel list.
  (4) Thus the random channel selection fails to find a suitable
      channel in the target BW and downgrades to a lower BW
      (40 MHz BW).
  (5) Now, with the target BW as 40 MHz, the only available
      channel in the primary channel list is still 52
      (20 MHz channel). The random channel selection fails
      to find a suitable channel in the target BW and
      downgrades to a lower BW (20 MHz BW) again.
  (6) Now, we have a 20 MHz channel available channel in the
      primary channel list (channel 52), so it selects channel
      52 in 20 MHz BW.

This is sub-optimal, as we still can operate with primary
channel as 52 and BW as 80

To fix this, enhance DFS random channel selection to include
all the channels that are available in the radio to build the
BW specific available channel list.

After a BW specific list is built and channels are available in
the target BW, intersect this with the primary channel list
and select a random channel from the intersected list.

If the target BW list is available and the intersected list
is empty, downgrade to a lower BW and try again.

Change-Id: Ibe883dd63092582e208f453ba088e0c669c7ca9b