qcacmn: clear osif_rx callback and osif peer
when ptqm migration happens we will delete the peer entries
becasue of this we won't find the peer in clearing the wds_ext dev,
so this peer will me maintain the stale vaules which will cause the
crash.
To avoid this crash we need find the peer entry by matching the osif peer pointer and clear wds_ext obejcts.

Change-Id: Idc7e3615b22a8bdf9e25111d1f823f26a732e397