qcacmn: Use AP type Vdev for starting agile RCAC
Currenlty, the WMI command to start the agile CAC
'WMI_VDEV_ADFS_CH_CFG_CMDID' is sent on the first active vdev
in the radio. But, in case of a repeater, the first vdev
in the vdev_list of the radio could be a STA Vap.

If the STA vdev is active, we end up sending the agile START
for a STA type vdev. This causes FW assert when FW validates
the ADFS and starts agile PRECAC.

The FW expects the vdev id:
 - Should be of an AP type vdev. (Or)
 - Should be of a STA type vdev only if STA DFS functionality is
   enabled.

Since both of these conditions are not satisfied, FW rejects the
ADFS config and asserts.

To fix this, make sure that host sends the vdev id only corresponding
to the AP type vdev in the radio.

Change-Id: I7767ffaff271d398f0f3c058a1b08c2b2d6ec16f