qcacmn: Fix Assert issue while peer cleanup
As part of peer unmap we delete all the ast peer
and a sanity check is failing due to intentional assert added
for lithium.

peer id for ast peer and peer id for actual peer for which unmap
is triggered is different causing the assert.

For BE this assert is not needed as WLAN FW does the AST addition,
and at HOST the sanity check for peer id while removing the peer can
be remove.

Change-Id: I9c90146ea78b9f76ef7d95d8559c76d4c6df42e4