qcacmn: Update phymode assignment for 6GHz channels for foreground scan
When wideband scan is enabled, Firmware expects 6GHz channels to have
an assigned phymode of 16 (11AX_HE20) during foreground scanning.

The current assignments set the phymode to 1 (11G) for all 2G
channels and 0 (11A) to every other channel, including the 6GHz
channels. Assigning the phymode based on center frequency and
differentiating between 5G and 6G channels will ensure that
host meets the expectations of Firmware.

Additionally, we do not see this issue when wideband scan is disabled
since the phymode check performed by Firmware is bypassed in this
case.

Change-Id: I0f8db31aaa76992041629828e96734f36cc7e364
CRs-Fixed: 4294251
Signed-off-by: Abhinav Raghunandan <abhiragh@qti.qualcomm.com>