qcacmn: Add an API to check if a given channel BW and PP is present
Add an API to check if a channel frequency, bandwidth and puncture
pattern is present in the AFC response payload information.

Also, write dispatchers for the API for use in other modules.

This API is used to check if a channel is available in AFC payload
in the change 'Ie113a447c0bd42997ec0fe07518f1b2a4944dc8d'

Change-Id: I80aee40bbbb0a30c010c0e81d0afdea47ff18fb2